backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
httpx==0.26.0

# Environment variables
python-dotenv==1.0.0

# CORS
python-multipart==0.0.6

# Document generation
python-docx==1.1.0
reportlab==4.0.7